Add Circle beside Marker to improve user experience

+++ starter/app/src/main/java/com/udacity/project4/locationreminders/savereminder/SaveReminderFragment.kt
@@ -258,7 +258,7 @@ class SaveReminderFragment : BaseFragment() {
             .setCircularRegion(
                 reminderDataItem.latitude!!,
                 reminderDataItem.longitude!!,
-                GeofencingConstants.GEOFENCE_RADIUS_IN_METERS
+                GeofencingConstants.GEOFENCE_RADIUS_IN_METERS.toFloat()
             )
             .setExpirationDuration(Geofence.NEVER_EXPIRE)
             .setTransitionTypes(Geofence.GEOFENCE_TRANSITION_ENTER)
